fix issue #97, issue #158, fix a plot typo, and happy 1000th commit

+++ subprojects/goojax/src/main/java/goojax/Cursor.java
@@ -5,9 +5,9 @@ package goojax;
 public class Cursor {
 	public String moreResultsUrl;
 	public int currentPageIndex;
-	public int estimatedResultCount;
+	public long estimatedResultCount;
 	public CursorPage pages[];
-	
+
 	Cursor() {}
 
 }
